Homework 2/requirements.txt
asgiref==3.9.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
fastjsonschema==2.21.2
fonttools==4.55.3
gunicorn==23.0.0
iniconfig==2.1.0
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
pillow==11.0.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.0
pytest==8.4.1
python-dateutil==2.9.0.post0
pywin32==311
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.11.0
